+++ src/main/java/io/streammachine/driver/client/SenderService.java
@@ -1,25 +1,28 @@
 package io.streammachine.driver.client;
 
+import io.streammachine.driver.common.CompletableFutureResponseListener;
 import io.streammachine.driver.domain.Config;
 import io.streammachine.driver.domain.StreamMachineEvent;
 import io.streammachine.driver.domain.StreamMachineEventDTO;
 import io.streammachine.driver.serializer.SerializationType;
-import org.asynchttpclient.AsyncHttpClient;
-import org.asynchttpclient.Response;
+import org.eclipse.jetty.client.HttpClient;
+import org.eclipse.jetty.client.api.ContentResponse;
+import org.eclipse.jetty.client.util.BytesContentProvider;
+import org.eclipse.jetty.http.HttpHeader;
+import org.eclipse.jetty.http2.client.HTTP2Client;
+import org.eclipse.jetty.http2.client.http.HttpClientTransportOverHTTP2;
+import org.eclipse.jetty.util.ssl.SslContextFactory;
 
 import java.util.concurrent.CompletableFuture;
 
-import static org.asynchttpclient.Dsl.asyncHttpClient;
-
 class SenderService {
     private final String endpointUri;
-    private final AsyncHttpClient client;
+    private final HttpClient httpClient;
     private final AuthService authService;
     private final Config config;
 
     public SenderService(String billingId, String clientId, String clientSecret, Config config) {
-        this.endpointUri = String.format("%s://%s%s", config.getGatewayScheme(), config.getGatewayHost(), config.getGatewayEndpoint());
-        this.client = asyncHttpClient();
+        this.endpointUri = String.format("%s://%s:%s%s", config.getGatewayScheme(), config.getGatewayHost(), config.getGatewayPort(), config.getGatewayEndpoint());
         this.authService = AuthService.builder()
                 .purpose(this.getClass().getSimpleName())
                 .billingId(billingId)
@@ -27,24 +30,36 @@ class SenderService {
                 .clientSecret(clientSecret)
                 .config(config)
                 .build();
-
         this.config = config;
+
+        HTTP2Client http2Client = new HTTP2Client();
+        SslContextFactory sslContextFactory = new SslContextFactory.Client();
+        http2Client.addBean(sslContextFactory);
+        httpClient = new HttpClient(new HttpClientTransportOverHTTP2(http2Client), sslContextFactory);
+
+        try {
+            http2Client.start();
+            httpClient.start();
+        } catch (Exception e) {
+            throw new IllegalStateException("An unexpected error occurred while starting a new Sender for Stream Machine.", e);
+        }
     }
 
-    public CompletableFuture<Response> send(StreamMachineEvent event, SerializationType type) {
+    public CompletableFuture<ContentResponse> send(StreamMachineEvent event, SerializationType type) {
         StreamMachineEventDTO dto = new StreamMachineEventDTO(event, type);
+        CompletableFuture<ContentResponse> completableFuture = new CompletableFuture<>();
 
-        return client.preparePost(endpointUri)
-                .setBody(dto.serialize())
-                .addHeader("Authorization", getBearerHeaderValue())
-                .addHeader("Content-Type", "application/octet-stream")
-                .addHeader("Strm-Serialization-Type", dto.getSerializationTypeHeader())
-// TODO retrieve version number from Maven / Gitlab CI build
-                .addHeader("Strm-Driver-Version", config.getImplementationVersion())
-                .addHeader("Strm-Driver-Build", "GET FROM VERSION FILE-Add with Maven")
-                .addHeader("Strm-Schema-Id", dto.getSchemaId())
-                .execute()
-                .toCompletableFuture();
+        httpClient.POST(this.endpointUri)
+                .header(HttpHeader.AUTHORIZATION, getBearerHeaderValue())
+                .header(HttpHeader.CONTENT_TYPE, "application/octet-stream")
+                .header("Strm-Driver-Version", config.getImplementationVersion())
+                .header("Strm-Driver-Build", "GET FROM VERSION FILE-Add with Maven")
+                .header("Strm-Serialization-Type", dto.getSerializationTypeHeader())
+                .header("Strm-Schema-Id", dto.getSchemaId())
+                .content(new BytesContentProvider(dto.serialize()))
+                .send(new CompletableFutureResponseListener(completableFuture));
+
+        return completableFuture;
     }
 
     private String getBearerHeaderValue() {