+++ src/main/java/io/streammachine/driver/client/StreamMachineClient.java
@@ -1,16 +1,14 @@
 package io.streammachine.driver.client;
 
+import io.streammachine.driver.common.WebSocketConsumer;
 import io.streammachine.driver.domain.Config;
 import io.streammachine.driver.domain.StreamMachineEvent;
 import io.streammachine.driver.serializer.SerializationType;
 import lombok.Builder;
 import lombok.extern.slf4j.Slf4j;
-import org.asynchttpclient.Response;
-import org.asynchttpclient.ws.WebSocketListener;
+import org.eclipse.jetty.client.api.ContentResponse;
 
-import javax.ws.rs.sse.InboundSseEvent;
 import java.util.concurrent.CompletableFuture;
-import java.util.function.Consumer;
 
 
 /**
@@ -32,25 +30,10 @@ public class StreamMachineClient {
      *
      * @param event the event.
      * @param type  the serialization type. only for Avro
-     * @return a future http response
+     * @return future containing the content response
      */
-    public CompletableFuture<Response> send(StreamMachineEvent event, SerializationType type) {
+    public CompletableFuture<ContentResponse> send(StreamMachineEvent event, SerializationType type) {
         return senderService.send(event, type);
-    }
-
-    /**
-     * Start an endless loop that receives events and applies `consumer` to them.
-     * Deprecated in favor of websocket
-     *
-     * @param asJson   tells the server to convert the event data to json. Otherwise you get base64 encoded data
-     *                 that might be avro binary, avro json or json depending on the schema and the serialization type.
-     *                 <p>
-     *                 This will be handled in a future version of the Java driver.
-     * @param consumer The consumer to apply to the sse
-     */
-    @Deprecated
-    public void startReceivingSse(boolean asJson, Consumer<InboundSseEvent> consumer) {
-        receiverService.receiveSse(asJson, consumer);
     }
 
     /**
@@ -62,7 +45,7 @@ public class StreamMachineClient {
      *                 This will be handled in a future version of the Java driver.
      * @param consumer The consumer to apply to the ws events
      */
-    public void startReceivingWs(boolean asJson, WebSocketListener consumer) {
+    public void startReceivingWs(boolean asJson, WebSocketConsumer consumer) {
         receiverService.receiveWs(asJson, consumer);
     }
 
@@ -71,7 +54,7 @@ public class StreamMachineClient {
      *
      * @return ok
      */
-    public Response egressIsAlive() {
+    public ContentResponse egressIsAlive() {
         return receiverService.isAlive();
     }
 }