+++ src/main/java/io/streammachine/driver/common/WebSocketConsumer.java
@@ -1,0 +1,13 @@
+package io.streammachine.driver.common;
+
+import org.eclipse.jetty.websocket.api.WebSocketAdapter;
+
+import java.util.concurrent.CountDownLatch;
+
+public class WebSocketConsumer extends WebSocketAdapter {
+    private final CountDownLatch closureLatch = new CountDownLatch(1);
+
+    public void awaitClosure() throws InterruptedException {
+        closureLatch.await();
+    }
+}